+++ app/(setup)/page.js
@@ -1,26 +1,16 @@
-import { initialProfile } from "@/lib/initial-profile";
-import { db } from "@/lib/db";
-import { redirect } from "next/navigation";
-import { InititalModal } from "@/components/modals/initial-modal";
+"use client";
+import { redirect } from 'next/navigation'
+import React from 'react'
+import { useRouter } from 'next/navigation';
 
-const SetupPage = async () => {
-  const profile = await initialProfile();
+const HomePage = () => {
+  const router = useRouter();
+  return (
+    <div>
+      <button className='bg-white text-black' 
+      onClick={()=> router.push("/user-setup")}>Sign-in</button>
+    </div>
+  )
+}
 
-  const server = await db.server.findFirst({
-    where: {
-      members: {
-        some: {
-          profileId: profile.id,
-        },
-      },
-    },
-  });
-
-  if (server) {
-    return redirect(`/servers/${server.id}`);
-  }
-
-  return <InititalModal />;
-};
-
-export default SetupPage;
+export default HomePage
