+++ app/(setup)/user-setup/page.js
@@ -1,0 +1,26 @@
+import { initialProfile } from "@/lib/initial-profile";
+import { db } from "@/lib/db";
+import { redirect } from "next/navigation";
+import { InititalModal } from "@/components/modals/initial-modal";
+
+const SetupPage = async () => {
+  const profile = await initialProfile();
+
+  const server = await db.server.findFirst({
+    where: {
+      members: {
+        some: {
+          profileId: profile.id,
+        },
+      },
+    },
+  });
+
+  if (server) {
+    return redirect(`/servers/${server.id}`);
+  }
+
+  return <InititalModal />;
+};
+
+export default SetupPage;
