+++ lib/initial-profile.js
@@ -1,13 +1,14 @@
 import { currentUser , redirectToSignIn } from "@clerk/nextjs";
 
 import {db} from '@/lib/db';
+import { redirect } from "@/node_modules/next/navigation";
 
 export const initialProfile = async ()=>{
 
     const user = await currentUser();
 
     if(!user) {
-        return redirectToSignIn();
+        return redirect("/sign-in");
     }
 
     const profile = await db.profile.findUnique({
